Core file reading functionality
Most of the core functionality for reading and analyzing files has been implemented. The program will now react to file changes, and is able to pick up and analyze some HTML tags. Only working as a proof of concept, a lot of content is still under development.

Currently requires an HTML folder to be on the same directory level as the .jar