Core Functionality
Almost all the core functionality of the program has now been implemented.

It is now possible to add new elements, edit new elements, as well as edit the CSS styles by tag, ID or class. 

Most of the UI functionality has been implemented. It is possible to drag and drop to re-arrange HTML elements, and all changes are reflected dynamically on the preview pane of the program. 

There are a few missing functionalities, such as adding a new file or a folder to the project, as well as creating CSS new classes.